+++ frontend/src/components/layout/box/box.tsx
