+++ frontend/src/components/layout/box/index.ts
@@ -1,0 +1,1 @@
+export { Box, type BoxProps } from './box'
