+++ frontend/src/components/layout/flex/index.ts
@@ -1,0 +1,1 @@
+export { Flex, type FlexProps } from './flex'
