+++ frontend/src/components/layout/grid/index.ts
@@ -1,0 +1,1 @@
+export { Grid, type GridProps } from './grid'
