+++ frontend/src/components/typo/heading/index.ts
@@ -1,0 +1,1 @@
+export { Heading, type HeadingProps } from './heading'
