+++ frontend/src/components/typo/text/index.ts
@@ -1,0 +1,1 @@
+export { Text, type TextProps } from './text'
