+++ frontend/src/components/ui/button/index.ts
@@ -1,0 +1,1 @@
+export { Button, type ButtonProps } from './button'
