+++ frontend/src/emotion.d.ts
@@ -1,7 +1,7 @@
 /// <reference types="@emotion/react/types/css-prop" />
 
 import '@emotion/react'
-import { Theme as MyTheme } from './styles/theme'
+import type { MyTheme } from './styles/theme'
 
 declare module '@emotion/react' {
   export interface Theme extends MyTheme {}
